chore: move and rename models.ts
Now /api/resources.ts
Drop .types from requests and responses files
