+++ frontend/src/types/index.ts
@@ -1,12 +1,12 @@
-export * from '@/types/common.ts';
-export * from '@/types/intakes.ts';
-export * from '@/types/models.ts';
-export * from '@/types/oidc.ts';
-export * from '@/types/ui.ts';
-export * from '@/types/validation.ts';
-export * from '@/types/api/dto.ts';
-export * from '@/types/api/request.types.ts';
-export * from '@/types/api/response.types.ts';
-export * from '@/types/api/external/ats.ts';
-export * from '@/types/api/external/geocoder.ts';
-export * from '@/types/api/external/orgBook.ts';
+export * from './common.ts';
+export * from './intakes.ts';
+export * from './oidc.ts';
+export * from './ui.ts';
+export * from './validation.ts';
+export * from './api/dto.ts';
+export * from './api/requests';
+export * from './api/resources';
+export * from './api/responses';
+export * from './api/external/ats.ts';
+export * from './api/external/geocoder.ts';
+export * from './api/external/orgBook.ts';
